Replaced session storage with database

+++ internal/configuration/dataStorage/DataStorage.go
@@ -7,6 +7,7 @@ import (
 	"encoding/binary"
 	"encoding/gob"
 	"errors"
+	"fmt"
 	"git.mills.io/prologic/bitcask"
 	"log"
 	"strings"
@@ -14,6 +15,7 @@ import (
 )
 
 const prefixFile = "file:id:"
+const prefixSessions = "session:id:"
 const idDefaultDownloads = "default:downloads"
 const idDefaultExpiry = "default:expiry"
 const idDefaultPassword = "default:password"
@@ -26,6 +28,37 @@ func Init(dbPath string) {
 		log.Fatal(err)
 	}
 	database = db
+}
+
+func GetSession(id string) (models.Session, bool) {
+	result := models.Session{}
+	if !database.Has([]byte(prefixSessions + id)) {
+		return result, false
+	}
+	value, err := database.Get([]byte(prefixSessions + id))
+	helper.Check(err)
+	buf := bytes.NewBuffer(value)
+	dec := gob.NewDecoder(buf)
+	err = dec.Decode(&result)
+	helper.Check(err)
+	return result, true
+}
+
+func DeleteSession(id string) {
+	if !database.Has([]byte(prefixSessions + id)) {
+		return
+	}
+	err := database.Delete([]byte(prefixSessions + id))
+	helper.Check(err)
+}
+
+func SaveSession(id string, session models.Session, expiry time.Duration) {
+	var buf bytes.Buffer
+	enc := gob.NewEncoder(&buf)
+	err := enc.Encode(session)
+	helper.Check(err)
+	err = database.PutWithTTL([]byte(prefixSessions+id), buf.Bytes(), expiry)
+	helper.Check(err)
 }
 
 func GetAllFiles() map[string]models.File {
@@ -67,6 +100,9 @@ func SaveMetaData(file models.File, expiry time.Duration) {
 }
 
 func DeleteMetaData(file models.File) {
+	if !database.Has([]byte(prefixFile + file.Id)) {
+		return
+	}
 	err := database.Delete([]byte(prefixFile + file.Id))
 	helper.Check(err)
 }
@@ -78,18 +114,12 @@ func GetUploadDefaults() (int, int, string) {
 	if database.Has([]byte(idDefaultDownloads)) {
 		bufByte, err := database.Get([]byte(idDefaultDownloads))
 		helper.Check(err)
-		var bufInt uint32
-		err = binary.Read(bytes.NewBuffer(bufByte), binary.LittleEndian, &bufInt)
-		helper.Check(err)
-		downloads = int(bufInt)
+		downloads = byteToInt(bufByte)
 	}
 	if database.Has([]byte(idDefaultExpiry)) {
 		bufByte, err := database.Get([]byte(idDefaultExpiry))
 		helper.Check(err)
-		var bufInt uint32
-		err = binary.Read(bytes.NewBuffer(bufByte), binary.LittleEndian, &bufInt)
-		helper.Check(err)
-		expiry = int(bufInt)
+		expiry = byteToInt(bufByte)
 	}
 	if database.Has([]byte(idDefaultPassword)) {
 		buf, err := database.Get([]byte(idDefaultPassword))
@@ -100,21 +130,32 @@ func GetUploadDefaults() (int, int, string) {
 }
 
 func SaveUploadDefaults(downloads, expiry int, password string) {
-	buf := make([]byte, 4)
-	binary.LittleEndian.PutUint32(buf, uint32(downloads))
-	err := database.Put([]byte(idDefaultDownloads), buf)
+	err := database.Put([]byte(idDefaultDownloads), intToByte(downloads))
 	helper.Check(err)
-
-	binary.LittleEndian.PutUint32(buf, uint32(expiry))
-	err = database.Put([]byte(idDefaultExpiry), buf)
+	err = database.Put([]byte(idDefaultExpiry), intToByte(expiry))
 	helper.Check(err)
-
 	err = database.Put([]byte(idDefaultPassword), []byte(password))
 	helper.Check(err)
 }
 
 func Close() {
 	if database != nil {
-		database.Close()
+		err := database.Close()
+		if err != nil {
+			fmt.Println(err)
+		}
 	}
+}
+
+func intToByte(integer int) []byte {
+	result := make([]byte, 4)
+	binary.LittleEndian.PutUint32(result, uint32(integer))
+	return result
+}
+
+func byteToInt(intByte []byte) int {
+	var result uint32
+	err := binary.Read(bytes.NewBuffer(intByte), binary.LittleEndian, &result)
+	helper.Check(err)
+	return int(result)
 }
